contract updates
