+++ abis/V2Moloch.json
@@ -50,11 +50,89 @@
 				"internalType": "uint256",
 				"name": "_summoningTermination",
 				"type": "uint256"
+			},
+			{
+				"internalType": "bytes32",
+				"name": "_manifesto",
+				"type": "bytes32"
 			}
 		],
 		"payable": false,
 		"stateMutability": "nonpayable",
 		"type": "constructor"
+	},
+	{
+		"anonymous": false,
+		"inputs": [
+			{
+				"indexed": true,
+				"internalType": "address",
+				"name": "depositToken",
+				"type": "address"
+			},
+			{
+				"indexed": true,
+				"internalType": "address",
+				"name": "minion",
+				"type": "address"
+			},
+			{
+				"indexed": false,
+				"internalType": "uint256",
+				"name": "periodDuration",
+				"type": "uint256"
+			},
+			{
+				"indexed": false,
+				"internalType": "uint256",
+				"name": "votingPeriodLength",
+				"type": "uint256"
+			},
+			{
+				"indexed": false,
+				"internalType": "uint256",
+				"name": "gracePeriodLength",
+				"type": "uint256"
+			},
+			{
+				"indexed": false,
+				"internalType": "uint256",
+				"name": "proposalDeposit",
+				"type": "uint256"
+			},
+			{
+				"indexed": false,
+				"internalType": "uint256",
+				"name": "dilutionBound",
+				"type": "uint256"
+			},
+			{
+				"indexed": false,
+				"internalType": "uint256",
+				"name": "processingReward",
+				"type": "uint256"
+			},
+			{
+				"indexed": false,
+				"internalType": "uint256",
+				"name": "summoningRate",
+				"type": "uint256"
+			},
+			{
+				"indexed": false,
+				"internalType": "uint256",
+				"name": "summoningTermination",
+				"type": "uint256"
+			},
+			{
+				"indexed": false,
+				"internalType": "bytes32",
+				"name": "manifesto",
+				"type": "bytes32"
+			}
+		],
+		"name": "AmendGovernance",
+		"type": "event"
 	},
 	{
 		"anonymous": false,
@@ -284,9 +362,9 @@
 			},
 			{
 				"indexed": false,
-				"internalType": "string",
+				"internalType": "bytes32",
 				"name": "details",
-				"type": "string"
+				"type": "bytes32"
 			},
 			{
 				"indexed": false,
@@ -409,6 +487,24 @@
 				"internalType": "uint256",
 				"name": "processingReward",
 				"type": "uint256"
+			},
+			{
+				"indexed": false,
+				"internalType": "uint256",
+				"name": "summoningRate",
+				"type": "uint256"
+			},
+			{
+				"indexed": false,
+				"internalType": "uint256",
+				"name": "summoningTermination",
+				"type": "uint256"
+			},
+			{
+				"indexed": false,
+				"internalType": "bytes32",
+				"name": "manifesto",
+				"type": "bytes32"
 			}
 		],
 		"name": "SummonComplete",
@@ -520,6 +616,71 @@
 		],
 		"payable": false,
 		"stateMutability": "view",
+		"type": "function"
+	},
+	{
+		"constant": false,
+		"inputs": [
+			{
+				"internalType": "address",
+				"name": "_depositToken",
+				"type": "address"
+			},
+			{
+				"internalType": "address",
+				"name": "_minion",
+				"type": "address"
+			},
+			{
+				"internalType": "uint256",
+				"name": "_periodDuration",
+				"type": "uint256"
+			},
+			{
+				"internalType": "uint256",
+				"name": "_votingPeriodLength",
+				"type": "uint256"
+			},
+			{
+				"internalType": "uint256",
+				"name": "_gracePeriodLength",
+				"type": "uint256"
+			},
+			{
+				"internalType": "uint256",
+				"name": "_proposalDeposit",
+				"type": "uint256"
+			},
+			{
+				"internalType": "uint256",
+				"name": "_dilutionBound",
+				"type": "uint256"
+			},
+			{
+				"internalType": "uint256",
+				"name": "_processingReward",
+				"type": "uint256"
+			},
+			{
+				"internalType": "uint256",
+				"name": "_summoningRate",
+				"type": "uint256"
+			},
+			{
+				"internalType": "uint256",
+				"name": "_summoningTermination",
+				"type": "uint256"
+			},
+			{
+				"internalType": "bytes32",
+				"name": "_manifesto",
+				"type": "bytes32"
+			}
+		],
+		"name": "amendGovernance",
+		"outputs": [],
+		"payable": false,
+		"stateMutability": "nonpayable",
 		"type": "function"
 	},
 	{
@@ -795,6 +956,21 @@
 	},
 	{
 		"constant": true,
+		"inputs": [],
+		"name": "manifesto",
+		"outputs": [
+			{
+				"internalType": "bytes32",
+				"name": "",
+				"type": "bytes32"
+			}
+		],
+		"payable": false,
+		"stateMutability": "view",
+		"type": "function"
+	},
+	{
+		"constant": true,
 		"inputs": [
 			{
 				"internalType": "address",
@@ -854,6 +1030,21 @@
 				"internalType": "uint256",
 				"name": "jailed",
 				"type": "uint256"
+			}
+		],
+		"payable": false,
+		"stateMutability": "view",
+		"type": "function"
+	},
+	{
+		"constant": true,
+		"inputs": [],
+		"name": "minion",
+		"outputs": [
+			{
+				"internalType": "address",
+				"name": "",
+				"type": "address"
 			}
 		],
 		"payable": false,
@@ -1058,9 +1249,9 @@
 				"type": "uint256"
 			},
 			{
-				"internalType": "string",
+				"internalType": "bytes32",
 				"name": "details",
-				"type": "string"
+				"type": "bytes32"
 			},
 			{
 				"internalType": "uint256",
@@ -1153,6 +1344,21 @@
 		"constant": false,
 		"inputs": [
 			{
+				"internalType": "address",
+				"name": "_minion",
+				"type": "address"
+			}
+		],
+		"name": "setMinion",
+		"outputs": [],
+		"payable": false,
+		"stateMutability": "nonpayable",
+		"type": "function"
+	},
+	{
+		"constant": false,
+		"inputs": [
+			{
 				"internalType": "uint256",
 				"name": "proposalId",
 				"type": "uint256"
@@ -1173,9 +1379,9 @@
 				"type": "address"
 			},
 			{
-				"internalType": "string",
+				"internalType": "bytes32",
 				"name": "details",
-				"type": "string"
+				"type": "bytes32"
 			}
 		],
 		"name": "submitGuildKickProposal",
@@ -1229,9 +1435,9 @@
 				"type": "address"
 			},
 			{
-				"internalType": "string",
+				"internalType": "bytes32",
 				"name": "details",
-				"type": "string"
+				"type": "bytes32"
 			}
 		],
 		"name": "submitProposal",
@@ -1275,9 +1481,9 @@
 				"type": "address"
 			},
 			{
-				"internalType": "string",
+				"internalType": "bytes32",
 				"name": "details",
-				"type": "string"
+				"type": "bytes32"
 			}
 		],
 		"name": "submitWhitelistProposal",
